Create My token_exchange.py
Here is a suggested GitHub page description for this code:

Title: MyToken Exchange

Description: This repository contains code for a simple decentralized exchange built on top of the Ethereum blockchain. The exchange allows users to trade MyToken, a custom ERC-20 token, using a web interface built with Flask. The code includes a Python script for interacting with the Ethereum blockchain using the Web3 library, a Solidity smart contract for the MyToken token, and a Flask web application for the exchange. 

Features:
- Connects to a local Ethereum node using Web3
- Deploys a MyToken smart contract on the Ethereum blockchain
- Allows users to view the balance of the MyToken contract address
- Allows users to transfer MyToken to other Ethereum addresses using the web interface
- Built with Python and Flask

This code can be used as a starting point for building other decentralized exchanges or for learning how to interact with the Ethereum blockchain using Python and Web3.